Fixed Syntax errors, but still does not successfully submit in current state
I fixed all syntax errors relating to "input" needing a string input to
be specified in some cases (was the root of the issue I've been having
with this all along".  Finally ran into a weird error that I saved in a
txt file and added to the repo under Testing.  Also a note: I tested my
parfor loop with the simulink simulation on my local computer using my
cores as workers, and it completed successfully. So feel free to use my
files to test, no configuration needed.  My input function is
paralleltestv2, all files are located in test_scripts.