Cleaned up
Cleaned up the old job-specific folders from friday.  The folders
currently are created during the config process but for some reason it
is not creating the job-specific environment file, so running
"cleanupJob" does not work due to not finding that file.  Likely due to
the process not finishing fully due to errors.